fix: PR fixes for toggle row
